requirements.txt
anyio==4.11.0
babel==2.17.0
backrefs==5.9
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
csscompressor==0.9.5
ghp-import==2.1.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
htmlmin2==0.1.13
httpcore==1.0.9
httpx==0.27.2
idna==3.10
Jinja2==3.1.6
jsmin==3.0.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-git-committers-plugin-2==2.5.0
mkdocs-git-revision-date-localized-plugin==1.4.7
mkdocs-material==9.6.21
mkdocs-material-extensions==1.3.1
mkdocs-minify-plugin==0.8.0
mkdocs_puml==2.3.0
msgpack==1.1.1
packaging==25.0
paginate==0.5.7
pathspec==0.12.1
platformdirs==4.4.0
Pygments==2.19.2
pymdown-extensions==10.16.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyyaml_env_tag==1.1
requests==2.32.5
rich==13.9.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
urllib3==2.5.0
watchdog==6.0.0

# Django Framework
Django==5.2

# API e documentação
djangorestframework==3.15.2
drf-yasg==1.21.7

# CORS
django-cors-headers==4.4.0

# PostgreSQL Database Adapter
psycopg2-binary==2.9.10

# Heroku Deployment
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.2.0
python-decouple==3.8
